2020-plant-explant/network_segment/requirements.txt
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.2
tensorboard==1.14.0
Pillow==6.2.0
opencv-python==4.1.0.25
future
scipy
scikit-image
